year-02/ca298/lab12/Lab_12/backend/requirements.txt
altgraph==0.17.4
appdirs==1.4.4
apturl==0.5.2
asgiref==3.8.1
attrs==22.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Bottleneck==1.3.2
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
cliapp==1.20180812.1
click==8.0.3
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.3
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.1
drf-spectacular-sidecar==2023.3.1
duplicity==0.8.21
et-xmlfile==1.0.1
fasteners==0.14.1
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
iniconfig==1.1.1
jdcal==1.0
jeepney==0.7.1
Jinja2==3.0.3
jsonschema==4.17.3
keyboard==0.13.5
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llvmlite==0.38.0
lockfile==0.12.2
loguru==0.7.0
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.3
MarkupSafe==2.0.1
matplotlib==3.5.1
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
mpmath==0.0.0
names==0.3.0
netifaces==0.11.0
numba==0.55.1
numexpr==2.8.1
numpy==1.24.2
oauthlib==3.2.0
odfpy==1.4.2
olefile==0.46
openpyxl==3.0.9
packaging==23.2
pandas==1.5.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==10.1.0
pluggy==0.13.0
ply==3.11
protobuf==3.12.4
ptyprocess==0.7.0
py==1.10.0
PyAutoGUI==0.9.54
pycairo==1.20.1
pycups==2.0.1
PyGetWindow==0.0.9
Pygments==2.11.2
PyGObject==3.42.1
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.11
PyJWT==2.3.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.19.3
PyScreeze==0.1.30
pytest==6.2.5
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python3-xlib==0.15
pythran==0.10.0
pytweening==1.0.7
pytz==2023.2
pyxdg==0.27
PyYAML==6.0
reportlab==3.6.8
requests==2.25.1
scipy==1.8.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.4
sympy==1.9
systemd-python==234
tables==3.7.0
toml==0.10.2
tqdm==4.66.1
ttystatus==0.38
typing_extensions==4.10.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
webencodings==0.5.1
wincertstore==0.2
xdg==5
xkit==0.0.0
xlwt==1.3.0
zipp==1.0.0
gunicorn==20.1.0